Pull out AsyncShardAction into its own class
